add multiplier from prstige

+++ hp2/agents/waiting_agent.py
@@ -180,7 +180,7 @@ def _compute_recipe_price(
     # Sum all per-ingredient costs declared in the config
     total_cost = sum(ing.get("price", 0.0) for ing in ingredient_list)
     # Apply archetype multiplier and round to integer (server requires int > 0)
-    price = max(1, round(total_cost * GLOBAL_MULTIPLIER)) 
+    price = max(1, int(round(total_cost * GLOBAL_MULTIPLIER)))
     return price
 
 # ---------------------------------------------------------------------------
